requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.27.0
pyyaml==6.0.2
google-cloud-secret-manager>=2.16.0
